task_manager/requirements.txt
Django>= 3.0
djangorestframework
djangorestframework-simplejwt
streamlit
requests
matplotlib
pandas